requirements.txt
fastapi[standard]==0.124.4
starlette[full]==0.50.0
python-dotenv==1.2.1
uvicorn==0.38.0
prometheus_fastapi_instrumentator==7.1.0

# Dev
autopep8==2.1.0
isort==5.12.0
pylint==3.1.0

# Build
openapi-spec-validator==0.7.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0  

# ai agent
langgraph>=0.4.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
pydantic>=2.0.0
redis>=5.0.0
langfuse>=3.0.0
langchain

